fix aux function for rotation, created one integration test and the circuit that integrates all the chips

+++ halo_blake2b_primitives/src/auxiliar_functions.rs
@@ -123,6 +123,6 @@ pub fn xor_field_elements<F: PrimeField>(a: F, b: F) -> F {
 pub(crate) fn rotate_right_field_element<F: PrimeField>(value_to_rotate: F, rotation_degree: usize) -> F {
     let value_to_rotate = convert_to_u64(value_to_rotate);
     let rotation_degree = rotation_degree % 64;
-    let rotated_value = (value_to_rotate >> rotation_degree) | (value_to_rotate << (64 - rotation_degree));
-    F::from(rotated_value)
+    let rotated_value = ((value_to_rotate as u128) >> rotation_degree) | ((value_to_rotate as u128) << (64 - rotation_degree));
+    F::from(rotated_value as u64)
 }
